\name{countries}
\alias{countries}
\docType{data}
\title{
idCountries
}
\description{
List of Countries and areas in countries with the ID used by FishBase and other SeaAround US project databases
}
\usage{data(countries)}
\format{
  A data frame with 248 observations on the following 2 variables.
  \describe{
    \item{\code{idCountry}}{a numeric vector}
    \item{\code{Country}}{A vecton containing the country names}
  }
}
\source{
EEZ ids in Sea Around Us project
}
\references{
Sea Around Us project Project
}
\examples{
data(countries)
head(countries)
}
\keyword{datasets}
