\name{FishBase-package}
\alias{FishBase-package}
\alias{FishBase}
\docType{package}
\title{
Obtain data from FishBase database
}
\description{
Download and parse the html code from several FishBase web pages, and put them into tables as data.frame.
}
\details{
\tabular{ll}{
Package: \tab FishBase\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-10-01\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Jorge Cornejo-Donoso

Maintainer: Jorge Cornejo-Dononoso <jcornejo@lifesci.ucsb.edu>
}
\references{
FishBase reference
}

\keyword{ package }

